backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ── Database (async) ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
greenlet==3.0.3

# ── Validation & Settings ─────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ── Auth / Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Celery / Redis ────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.8
flower==2.0.1
kombu==5.3.7

# ── AI / Vision ───────────────────────────────────────────────────────────────
openai==1.51.2
pillow==11.0.0
httpx==0.27.2

# ── Payments ──────────────────────────────────────────────────────────────────
stripe==10.12.0

# ── Utils ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.2.0
